refactor: remove separeted file for components types
